requirements.txt
numpy~=1.19.5
pandas~=1.1.5
PyYAML~=5.4.1
matplotlib~=3.3.4
untangle~=1.1.1